+++ backend/models/admin.js
@@ -1,9 +1,10 @@
 
 
 const mongoose = require('mongoose');
+const bcrypt = require('bcryptjs');
 
-const adminSchema = new mongoose.Schema({
-    text: {
+const addProductSchema = new mongoose.Schema({
+    title: {
         type: String,
         required: true
     },
@@ -37,9 +38,36 @@ const adminSchema = new mongoose.Schema({
     }
 });
 
-const Detail = mongoose.model('Detail', adminSchema);
+const adminSchema = new mongoose.Schema({
+  email: {
+    type: String,
+    required: true,
+    unique: true,
+  },
+  password: {
+    type: String,
+    required: true,
+  },
+});
 
-module.exports = Detail;
+// Hash password before saving the admin
+adminSchema.pre('save', async function (next) {
+  if (!this.isModified('password')) {
+    return next();
+  }
+  const salt = await bcrypt.genSalt(10);
+  this.password = await bcrypt.hash(this.password, salt);
+  console.log(this.password);
+  next();
+});
+
+const Admin = mongoose.model('Admin', adminSchema);
+
+const Detail = mongoose.model('Detail', addProductSchema);
+
+
+
+module.exports = {Detail, Admin};
