+++ frontend/src/pages/Admin/AdminLogin.jsx
@@ -1,0 +1,70 @@
+// src/pages/admin/AdminLogin.jsx
+
+import React, { useState } from "react";
+
+const AdminLogin = () => {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState("");
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setError("");
+
+    const response = await fetch("http://localhost:3000/api/v1/admin/login", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({ email, password }),
+    });
+
+    const data = await response.json();
+
+    if (response.ok) {
+      // Store the JWT in local storage
+      localStorage.setItem("token", data.token);
+      // Redirect to admin dashboard or other route
+    } else {
+      setError(data.message || "Login failed");
+    }
+  };
+
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gray-100 p-6">
+      <div className="bg-white shadow-lg rounded-lg p-6 max-w-sm w-full">
+        <h2 className="text-2xl font-semibold text-center text-gray-700 mb-6">
+          Admin Login
+        </h2>
+        <form onSubmit={handleSubmit}>
+          <div className="mb-4">
+            <label className="block text-lg font-medium text-gray-700 mb-2">Email</label>
+            <input
+              type="email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none"
+              required
+            />
+          </div>
+          <div className="mb-4">
+            <label className="block text-lg font-medium text-gray-700 mb-2">Password</label>
+            <input
+              type="password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none"
+              required
+            />
+          </div>
+          {error && <p className="text-red-500 mb-4">{error}</p>}
+          <button className="w-full py-3 bg-blue-500 text-white rounded-lg hover:bg-blue-600 transition duration-300">
+            Login
+          </button>
+        </form>
+      </div>
+    </div>
+  );
+};
+
+export default AdminLogin;
